ids[multipart]: refactor ids multipart handler package
