+++ data-protocols/ids/ids-api-multipart-endpoint-v1/src/main/java/org/eclipse/dataspaceconnector/ids/api/multipart/IdsMultipartApiServiceExtension.java
@@ -24,17 +24,17 @@ import org.eclipse.dataspaceconnector.ids.api.multipart.controller.MultipartCont
 import org.eclipse.dataspaceconnector.ids.api.multipart.handler.ArtifactRequestHandler;
 import org.eclipse.dataspaceconnector.ids.api.multipart.handler.DescriptionHandler;
 import org.eclipse.dataspaceconnector.ids.api.multipart.handler.Handler;
-import org.eclipse.dataspaceconnector.ids.api.multipart.handler.contract.ContractAgreementHandler;
-import org.eclipse.dataspaceconnector.ids.api.multipart.handler.contract.ContractOfferHandler;
-import org.eclipse.dataspaceconnector.ids.api.multipart.handler.contract.ContractRejectionHandler;
-import org.eclipse.dataspaceconnector.ids.api.multipart.handler.contract.ContractRequestHandler;
+import org.eclipse.dataspaceconnector.ids.api.multipart.handler.ContractAgreementHandler;
+import org.eclipse.dataspaceconnector.ids.api.multipart.handler.ContractOfferHandler;
+import org.eclipse.dataspaceconnector.ids.api.multipart.handler.ContractRejectionHandler;
+import org.eclipse.dataspaceconnector.ids.api.multipart.handler.ContractRequestHandler;
 import org.eclipse.dataspaceconnector.ids.api.multipart.handler.description.ArtifactDescriptionRequestHandler;
 import org.eclipse.dataspaceconnector.ids.api.multipart.handler.description.ConnectorDescriptionRequestHandler;
 import org.eclipse.dataspaceconnector.ids.api.multipart.handler.description.DataCatalogDescriptionRequestHandler;
 import org.eclipse.dataspaceconnector.ids.api.multipart.handler.description.RepresentationDescriptionRequestHandler;
 import org.eclipse.dataspaceconnector.ids.api.multipart.handler.description.ResourceDescriptionRequestHandler;
-import org.eclipse.dataspaceconnector.ids.api.multipart.handler.notifications.MessageProcessedNotificationHandler;
-import org.eclipse.dataspaceconnector.ids.api.multipart.handler.notifications.RequestInProcessHandler;
+import org.eclipse.dataspaceconnector.ids.api.multipart.handler.MessageProcessedNotificationHandler;
+import org.eclipse.dataspaceconnector.ids.api.multipart.handler.RequestInProcessHandler;
 import org.eclipse.dataspaceconnector.ids.spi.IdsId;
 import org.eclipse.dataspaceconnector.ids.spi.IdsIdParser;
 import org.eclipse.dataspaceconnector.ids.spi.IdsType;
