+++ data-protocols/ids/ids-api-multipart-endpoint-v1/src/main/java/org/eclipse/dataspaceconnector/ids/api/multipart/handler/ResponseMessageUtil.java
@@ -31,7 +31,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.UUID;
 
-public class ResponseMessageUtil {
+class ResponseMessageUtil {
 
     public static ResponseMessage createDummyResponse(
             @Nullable String connectorId,
